Final Commit w/ roles - Also In Readme.md
Abhijith Varma Mudunuri - Worked on Frontend and Local Backend primarily. Created Base template pages for my team to use in order to work on their respective content. Pushed all files for my team in order to block any merge requests. Created small projects for the website; calculator, python emulator, snake game, car game, quiz js, contact form js. I also helped work on the PHP and notes although it wasn’t my primary job or primary work. I basically created every HTML page.

Gerald - Worked on adding notes to the html files for the chapter notes. Also, edited the javascript files to implement various questions for the quizzes. Helped troubleshoot certain problems with the html web pages and styling errors. Also worked on creating the contact form to make the website more interactive. I connected the local frontend (the contact.html form) to the local backend (MySQL & PHP) using XAMPP. I made sure to specify the action as “connectDB.php” in the form tag for contact.html to connect the html file to the PHP. Anything inputted into the contact form was able to store information into the database.

Ruizhi - Designed the looks of the website, linked all the javascript quizzes to their respective html files, chose all the photos for the website, decided the content of the website, typed up and formatted the notes and quizzes, worked on the python emulator and participated to set up the database, worked on the navigation bar, contributed in styling the website using javascript and css, and tested the car/snake games. I also created the different JS quizzes and helped embed the calculator to the website on the overlay.